Partial mount and DC test
While shunting some capacitors and setting to a couple of mA Output,
this test, sending -max to +max input DC generates correct LED display and correct output current.

The through hole for the 1N4148 is a little big small. This is going to be fixed.

Next: cut the circuit, mount the supplies, mount the logic,
build a Q&D input oscillator and test using a scope.