First test
R203 and R206 are 4.7R because I have some. The load resistance is 390R.
It looks like good as 400mV/4.7 with a 3 ratio transformer should be 28mA.
I still need to know the current I need.
The definitive version is going to be the whose have these adjustments (including C45 an C46 to fit the configuration).