New current measure Spice circuit
This utility is not a part of the project. It is not involved in the Makefile.
To measure the current while stimulating for real, there is a need of isolated measurement equipment.
Not everybody has got a handled oscilloscope.
The work around is to build a small circuit and use a handled voltmeter.
The circuit is a low value resistor (here 90.9R), a 4 diodes bridge (1N4148), a capacitor (1000uF) and a high value discharge resistor.
However the dropout voltage of the diodes disturbs the measurement.
This file is intended to find the current that produces the same output voltage (here 65mA).