Make the box milling easy
The box meca file contains now the full in and out lines.
Some elements of the PCB (relative to the milling and the assembly)
  have been copied and pasted to the location of the second board.
Some quotation lines are extended for the whose want to print the box mechanics
  and use the paper to mark the holes.

It has no impact on the PCB as both the pcb and box meca files
  have nothing to do in an order folder.
Nothing changed on the PCB itself nor its outline.