PCB and drawings improvements
The lines pointing the center of the plugs and the buttons has been enlarged.
The drill of the plugs and the buttons has been drawn on the box meca plane.
A second drawing has been mirrored.

It is irrelevant for people who use CNC software.
It is nice for people who place the picture under the bottom of the box,
  fold the sides and mark the center of the holes.