Input sound 8 channels option.
The PCB is designed to assemble a proprietary DB9 for 4 channels input.
This PCB mount DB9 is designed to fit the box.

An option is to assemble an 8 channels DB25.
There are standard cables for the balanced analog in or out,
  or for the AES/EBU digital 4 in 4 out.
However, the footprint is only intended to make the wirering easy.
The holes are face to face with the DB pins.